Reorganize Components, remove unused files
