+++ src/components/SheetCard.js
@@ -11,10 +11,6 @@ export default function SheetCard(props) {
                 <Flex>
                     <Box variant='characterBox'>
                         <Text variant='cardHeader'>{props.name}</Text>
-                        <Text variant='ingestionText'>Race</Text>
-                        <Text variant='modalText'>Elf</Text>
-                        <Text variant='ingestionText'>Level</Text>
-                        <Text variant='modalText'>12</Text>
                         <Button><Link to="/user/sheet">Select {props.name} </Link></Button>
                     </Box>
                 </Flex>
