+++ src/components/widgets/ExampleWidget.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, Button, Heading, Text } from 'rebass';
-import Widget from './Widget';
+import Widget from '../Widget';
 
 // Usage: <ExampleWidget>Hello World!</ExampleWidget>
 export default class ExampleWidget extends Widget {
