+++ src/components/widgets/parts/InventoryRow.js
